Move type declarations into separate directory
