+++ lib/chat_history.ts
@@ -1,9 +1,9 @@
 import { EventEmitter } from "https://deno.land/x/event@2.0.1/mod.ts";
 import { AgentAction, BaseChatMessage } from "https://esm.sh/v118/langchain@0.0.67/schema.js";
 import { ChatMessage } from "https://esm.sh/langchain/schema";
-import { ChatMessage as ChatMessageT, ChatRole } from "../chat.d.ts";
-import { ChatEvent } from "../chat_events.d.ts";
-import { ChatHistory, ChatHistoryEvents } from "../plugins.d.ts";
+import { ChatMessage as ChatMessageT, ChatRole } from "../types/chat.d.ts";
+import { ChatEvent } from "../types/chat_events.d.ts";
+import { ChatHistory, ChatHistoryEvents } from "../types/plugins.d.ts";
 
 export class InMemoryChatHistory implements ChatHistory {
   public readonly events = new EventEmitter<ChatHistoryEvents>();
