+++ lib/stream_transformers.ts
@@ -1,6 +1,6 @@
 import { readableStreamFromIterable } from "https://deno.land/std@0.184.0/streams/readable_stream_from_iterable.ts";
 import { BaseChatMessage } from "https://esm.sh/v118/langchain@0.0.67/schema.js";
-import { ParameterType, Parameters, ParsedCodeBlockTag } from "../plugins.d.ts";
+import { ParameterType, Parameters, ParsedCodeBlockTag } from "../types/plugins.d.ts";
 
 export function ProcessMessageContent<T>(decode: (input: ReadableStream<string>) => ReadableStream<T>): TransformStream<BaseChatMessage, [T, BaseChatMessage]> {
   return new TransformStream<BaseChatMessage, [T, BaseChatMessage]>({
